Refactoring: Fix a lot of nonsense and move BoardPoint
This commit does some large-scale refactoring. In particular it enforces the rule to create and
include headers AND NOT IMPLEMENTATION FILES.

The run script has been changed to always compile with the cpp files in src/, so that the
implementations can be loaded.

I have changed a lot of includes and created prototypes for existing functions.

I have isolated BoardPoint and the convenience functions I made for it in new header and
implementation files.

Due to a compiler bug, I was also forced to make the methods of BoardPoint non-constexpr. :(

@TimBeckmann @Pfege These changes are important to consider when working on the repository.
I highly recommend rebasing your active branches once this is merged into master.

Closes #34