Automatically read version information from repository
This commit changes the version configuration and build scripts so that the tagged version or
information about it's relation to the latest version is baked into the executable.

This makes manually updating the version number after each tagging unnecessary.
Versions of the software not built with this information now always show a "(version unknown)"
instead.